dj_pr/requirements.txt
asgiref==3.7.2
aws-psycopg2==1.3.8
azure-core==1.26.4
azure-functions==1.14.0
azure-identity==1.13.0
azure-storage-blob==12.16.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
configparser==6.0.0
cryptography==41.0.1
defusedxml==0.7.1
distlib==0.3.6
Django==3.0.14
django-cors-headers==4.2.0
django-environ==0.10.0
django-storages==1.12.3
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
filelock==3.12.2
httpie==3.2.2
idna==3.4
isodate==0.6.1
markdown-it-py==2.2.0
mdurl==0.1.2
msal==1.22.0
msal-extensions==1.0.0
multidict==6.0.4
mysqlclient==2.2.0
Pillow==9.5.0
pipenv==2023.7.23
platformdirs==3.8.1
portalocker==2.7.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.21
Pygments==2.15.1
PyJWT==2.7.0
pymongo==3.12.3
PyMySQL==1.1.0
PySocks==1.7.1
pytz==2023.3
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.4.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.2
urllib3==2.0.2
virtualenv==20.23.1
virtualenv-clone==0.5.7
